+++ layers/complex_matrix/src/factories.cpp
@@ -2,27 +2,27 @@
 #include "random.hpp"
 
 
-ComplexMatrixGroup ComplexMatrixGroup::random(double abs_max, size_t n, size_t p)
+ComplexMatrixGroup ComplexMatrixGroup::random(long double abs_max, size_t n, size_t p)
 {
     size_t len = (p-1)*n*n;
     std::vector<complex> result(len);
     for(size_t i=0; i<len; i++)
     {
-        double abs = random_generators::random_real() * abs_max;
-        double angle = random_generators::random_real() * M_PI * 2;
+        long double abs = random_generators::random_real() * abs_max;
+        long double angle = random_generators::random_real() * M_PI * 2;
         result[i] = std::polar(abs, angle);
     }
     return ComplexMatrixGroup(n, p, result);
 
 }
 
-ComplexMatrixGroup ComplexMatrixGroup::random_imag(double abs_max, size_t n, size_t p)
+ComplexMatrixGroup ComplexMatrixGroup::random_imag(long double abs_max, size_t n, size_t p)
 {
     size_t len = (p-1)*n*n;
     std::vector<complex> result(len);
     for(size_t i=0; i<len; i++)
     {
-        double x = (2*random_generators::random_real()-1) * abs_max;
+        long double x = (2*random_generators::random_real()-1) * abs_max;
         result[i] = complex(0, x);
     }
     return ComplexMatrixGroup(n, p, result);
